requirements.txt
streamlit
langchain
langchain-groq
pypdf
chromadb
python-dotenv
